Check that interstitial contains the good creative
Tested scenario:
- Given already prefetch valid bid
- Given MoPub response pointing to Criteo adapter
- When loading the MoPub interstitial
- And waiting for idle state
- And showing the interstitial
- And waiting for an activity to be opened
- Then mopub listener is notified for shown interstitial
- Then interstitial contains a web view with the ad creative

JIRA: EE-1179
