+++ mediation/src/androidTest/java/com/criteo/publisher/view/WebViewLookupExt.kt
@@ -1,0 +1,13 @@
+package com.criteo.publisher.view
+
+import android.view.View
+import java.util.concurrent.Future
+
+fun WebViewLookup.lookForNonEmptyHtmlContent(view: View): Future<String> {
+  // TODO move this in test-utils module
+  lookForWebViews(view).forEach {
+    it.waitUntilLoaded()
+  }
+
+  return lookForHtmlContent(view)
+}
